+++ resources/views/cms/pages/administration/create.blade.php
